+++ pybop/optimisers/base_pints_optimiser.py
@@ -263,6 +263,7 @@ class BasePintsOptimiser(BaseOptimiser):
                 _fs = [x[0] for x in fs] if self._needs_sensitivities else fs
                 self.log_update(
                     iterations=iteration,
+                    evaluations=evaluations,
                     x=xs,
                     x_best=self.optimiser.x_best(),
                     cost=_fs,